requirements.txt
Flask
flask-cors
Werkzeug
python-dotenv

langchain
langchain-google-genai
langchain-pinecone
langchain-cohere

pinecone
cohere
google-generativeai
google-ai-generativelanguage

guardrails-ai
guardrails-api-client

